// config/build/types/config.ts
export type BuildMode = 'development' | 'production';

export interface BuildPaths {
    entry: string;
    build: string;
    html: string;
    src: string;
    locales: string;
    buildLocales: string;
    faviconPath: string;
}

export interface BuildEnv {
    mode: BuildMode;
    port: number;
    apiURL: string;
}

export interface buildOptions {
    mode: BuildMode;
    paths: BuildPaths;
    isDev: boolean;
    port: number;
    apiURL: string;
    project: 'storybook' | 'frontend' | 'jest';
}